refactor [server]: move dtos folder to module folder root
