+++ server/src/modules/article/application/article.service.ts
@@ -4,10 +4,10 @@ import { Repository } from 'typeorm';
 
 import { Article } from '../domain/entities/article.entity';
 import { IArticleRepository } from '../domain/interfaces/article.repository.interface';
+import { CreateArticleDto } from '../dtos/create-article.dto';
+import { UpdateArticleDto } from '../dtos/update-article.dto';
 import { ArticleRepository } from '../repository/typeorm/repository/article.repository';
 import { ArticleSchema } from '../repository/typeorm/schema/article.schema';
-import { CreateArticleDto } from './dtos/create-article.dto';
-import { UpdateArticleDto } from './dtos/update-article.dto';
 
 @Injectable()
 export class ArticleService {
